refactor(repo): 拆分超长页面与设置文件
