+++ docs/DESIGN.md
@@ -164,12 +164,7 @@ _initApp()
 
 **当前状态**：已接入多套真实只读教务服务
 
-**当前状态说明**：
-- 体育部课外活动考勤：使用独立体育部查询密码，只读读取早操、课外活动、次数调整、体育长廊统计与明细
-- 第二课堂学分：通过 OA/CAS 登录态进入学工报表系统，只读读取第二课堂逐项得分明细，不将单项分值相加为总学分
-- 本专科教务摘要：通过 OA `bzkjw` 入口复用 CAS 会话，聚合展示个人信息、当前课表、成绩、考试、培养计划与培养计划完成情况，并识别开课检索、空闲教室入口状态
-- 教学评价等写入型功能仍保持关闭，不在应用内暴露入口
-- 受限服务查询前统一执行校园网 / VPN 前置检测，不可达时不提交密码、不刷新 OA 会话
+**当前状态说明**：已接入体育部考勤、第二课堂学分和本专科教务摘要三类只读服务。受限服务查询前统一执行校园网 / VPN 前置检测，不可达时不提交密码、不刷新 OA 会话；教学评价等写入型功能保持关闭。
 
 ### 4.2.1 课程表页面（CourseSchedulePage）
 
@@ -177,12 +172,7 @@ _initApp()
 
 **当前状态**：已实现独立课程表页
 
-**已实现能力**：
-- 作为主导航独立页面展示当前学期本专科教务课表
-- 展示课程名称、时间、地点、教师、周次等核心字段
-- 未配置 OA 密码、未连接校园网 / VPN 或课表不可用时显示明确状态提示
-- 支持和教务中心共用同一套本专科教务自动刷新配置
-- 仅展示只读课表信息，不提供选课、退课、调课等写入入口
+**已实现能力**：作为主导航独立页面展示当前学期课表，复用本专科教务自动刷新配置，并在凭据、校园网 / VPN 或课表不可用时显示明确状态。页面仅展示课程名称、时间、地点、教师和周次，不提供选课、退课、调课等写入入口。
 
 ### 4.3 信息中心（InfoPage）
 
@@ -217,16 +207,7 @@ _initApp()
 
 **当前状态**：已实现只读收信与协议校验
 
-**已实现能力**：
-- 使用设置页保存的学工号派生学校邮箱账号，并读取独立邮箱密码
-- 通过 IMAP `EXAMINE INBOX` + `BODY.PEEK[]` 只读读取最近邮件
-- 通过 POP `LIST` + `RETR` 只读读取最近邮件
-- 通过 SMTP `EHLO` + `AUTH` 校验登录状态，不发送邮件
-- 展示邮件标题、发件人、时间、正文摘要和正文快照
-
-**只读边界**：
-- 不提供发送、回复、转发、删除、移动、标记已读或修改文件夹入口
-- 邮箱系统不执行校园网 / VPN 前置检测
+**已实现能力**：使用学工号派生学校邮箱账号，通过 IMAP / POP 只读读取最近邮件，并通过 SMTP 仅校验登录状态。页面展示标题、发件人、时间、正文摘要和正文快照，不提供发送、回复、转发、删除、移动、标记已读或修改文件夹入口。
 
 ### 4.5 设置页（SettingsPage）
 
@@ -244,25 +225,13 @@ _initApp()
 
 #### 4.5.2 自动刷新设置
 
-- **校园网 / VPN 状态检测间隔**：默认 15 分钟，可关闭自动检测；关闭后仍可点击导航栏状态徽标手动检测
-- **体育查询自动刷新**：默认关闭，可配置教务中心课外活动考勤卡片的自动读取间隔；关闭后仍可在卡片右下角手动刷新并查看上次刷新时间
-- **校园卡余额自动刷新**：默认关闭，可配置主页校园卡余额卡片的自动读取间隔；关闭后仍可在卡片右下角手动刷新并查看上次刷新时间
-- **学校邮箱自动刷新**：默认关闭，可配置学校邮箱页面的自动收信间隔；关闭后仍可在邮箱页手动读取最近邮件
-- **第二课堂学分自动刷新**：默认关闭，可配置教务中心第二课堂学分卡片的自动读取间隔；关闭后仍可在卡片右上角手动刷新并查看上次刷新时间
-- **本专科教务自动刷新**：默认关闭，可配置教务中心本专科教务摘要与独立课程表页面的自动读取间隔；关闭后仍可在对应页面手动刷新
-- **自动刷新快捷入口**：提供职能部门、教学单位、微信推文三个入口，跳转到对应分区顶部的自动刷新设置面板
+集中配置校园网 / VPN 状态检测、体育查询、校园卡余额、学校邮箱、第二课堂学分和本专科教务的自动刷新开关与间隔。默认不自动访问受限服务，关闭后仍可在对应页面手动刷新；职能部门、教学单位、微信推文保留快捷入口跳转到对应设置面板。
 
 #### 4.5.3 安全设置
 
-- **密码保护开关**：`ToggleSwitch` 控制启用/禁用
-  - 开启时弹出"设置密码"对话框
-  - 关闭时弹出"移除密码"对话框（需验证当前密码）
-- **修改密码**：仅在已设置密码时显示，需验证旧密码
-- **系统快速验证**：仅在密码保护已开启且 Android / iOS / macOS / Windows 设备支持系统认证时显示开关；不可用或未配置时显示密码兜底提示，不改变手动密码解锁语义
-- **立即上锁**：不退出应用，直接回到锁定页
-- **教务凭据**：保存学工号、OA 密码、体育部查询密码和邮箱密码，学校邮箱账号固定由学工号派生；密码框回访时保持为空，仅显示“已填写 / 未填写”
-- **清理信息中心缓存**：只清理消息缓存与已读状态
-- **清除所有本地数据**：清空状态文件、教务安全存储项并退出应用
+- **密码保护**：支持启用、移除、修改密码、立即上锁和系统快速验证兜底提示
+- **教务凭据**：保存学工号、OA 密码、体育部查询密码和邮箱密码；学校邮箱账号固定由学工号派生
+- **数据清理**：支持清理信息中心缓存或清除所有本地数据
 
 #### 4.5.4 消息推送设置
 
@@ -506,44 +475,13 @@ ContentDialog: 输入当前密码
 
 ## 8. 目录结构
 
-```
-lib/
-├── main.dart                         # 应用入口、平台初始化、EULA/锁定流程
-├── app.dart                          # 桌面导航与移动端底部导航
-├── controllers/
-│   └── settings_wechat_controller.dart
-├── models/                           # 渠道配置、消息模型、微信矩阵模型
-├── pages/
-│   ├── home_page.dart
-│   ├── academic_page.dart
-│   ├── course_schedule_page.dart
-│   ├── info_page.dart
-│   ├── quick_links_page.dart
-│   ├── settings_page.dart
-│   ├── about_page.dart
-│   ├── lock_page.dart
-│   ├── webview_page.dart
-│   └── wxmp_login_page.dart
-├── services/                         # 持久化、抓取、自动刷新、通知、退出、版本信息
-├── theme/                            # Fluent 2 token 体系
-├── utils/                            # WebView 环境、时间工具、微信匹配工具
-└── widgets/                          # 设置分区、频道列表、响应式与消息项组件
-```
+核心目录包括 `lib/main.dart`、`lib/app.dart`、`lib/controllers/`、`lib/models/`、`lib/pages/`、`lib/services/`、`lib/theme/`、`lib/utils/` 和 `lib/widgets/`。页面、服务和模型按功能继续拆分，超长文件默认继续解耦。
 
 ---
 
 ## 9. 待实现功能清单
 
-| 优先级 | 功能 | 涉及页面 | 依赖 |
-|--------|------|----------|------|
-| P0 | 教务真实接口接入 | AcademicPage | 教务系统鉴权与接口/网页抓取 |
-| P0 | Android / 桌面 Release 持续集成完善 | 构建流程 | GitHub Actions / 平台签名材料 |
-| P1 | 信息中心抓取源继续扩充 | InfoPage | 新站点解析规则 |
-| P1 | 主页卡片扩展为可配置摘要面板 | HomePage | 本地状态与组件抽象 |
-| P1 | 快速跳转支持用户自定义与排序 | QuickLinksPage | 本地配置写回 |
-| P2 | 多语言支持 | 全局 | flutter_localizations |
-| P2 | 数据导出/备份 | SettingsPage | 文件系统 |
-| P2 | 发布安装器与签名公证完善 | 各平台 | 平台证书与打包工具 |
+后续重点包括 Android / 桌面 Release 持续集成完善、信息中心抓取源扩充、主页可配置摘要面板、快速跳转用户自定义与排序、多语言支持、数据导出 / 备份，以及安装器与签名公证完善。
 
 ---
