+++ docs/USAGE.md
@@ -171,169 +171,39 @@ flutter test --coverage
 
 ## 6.5 本专科教务 / OA 登录校验调试
 
-安全设置页的“教务凭据”区域提供“验证 OA 登录”按钮，用于只读检查已保存的 OA 账号密码是否可完成本专科教务系统入口登录。
+安全设置页的“教务凭据”区域提供“验证 OA 登录”按钮，用于只读检查已保存的 OA 账号密码是否可完成学校 OA / CAS 登录。校验前会先执行校园网 / VPN 前置检测；成功后仅保存可复用 Cookie 会话快照，不保存一次性 CAS Ticket。
 
-校验流程：
-
-- 先执行校园网 / VPN 前置检测；不可访问受限校园站点时不提交 OA 凭据
-- 访问 `https://oa.sspu.edu.cn/interface/Entrance.jsp?id=bzkjw`，按网页跳转进入 `id.sspu.edu.cn/cas/login`
-- 从 CAS 登录页读取 `execution` 等隐藏字段，并通过 `/cas/jwt/publicKey` 获取 RSA 公钥
-- 按 CAS 页面规则提交 `__RSA__` 前缀的 RSA 加密密码；成功条件是 CAS 跳转回 `oa.sspu.edu.cn`
-- 登录成功后仅将响应链路中的 Cookie 会话快照写入系统安全存储，供后续只读网页请求复用；不保存一次性 CAS Ticket
-
-已知交互限制：
-
-- CAS 要求图形验证码时，应用会提示“需要图形验证码”，不会尝试绕过验证码
-- CAS 要求 MFA / 安全验证时，应用会提示“需要额外安全验证”，不会继续无交互登录
-- CAS 登录页结构缺少表单、`execution` 或公钥解析失败时，按网页流程异常处理
-- OA 登录持久性较低；后续请求发现跳回 CAS 或返回未登录状态时，可直接重新点击“验证 OA 登录”刷新会话
-- 登录校验仅允许登录状态判断，不执行查询、提交、标记、删除等写入或状态变更操作
-
-登录规则探索记录见 [OA_LOGIN_RULES.md](OA_LOGIN_RULES.md)。
+CAS 要求图形验证码、MFA / 安全验证或页面结构变化时，应用会给出明确状态提示，不会尝试绕过交互验证。登录规则探索记录见 [OA_LOGIN_RULES.md](OA_LOGIN_RULES.md)。
 
 ## 6.6 体育部课外活动考勤调试
 
-教务中心页提供体育部查询系统的课外活动考勤卡片，展示总次数和四类拆分：早操、课外活动、次数调整、体育长廊。
+教务中心页提供体育部查询系统的课外活动考勤卡片，展示总次数和早操、课外活动、次数调整、体育长廊四类明细。默认不自动访问体育部系统；可手动刷新，也可在设置页开启“体育查询自动刷新”。
 
-刷新方式：
-
-- 默认不自动访问体育部查询系统，避免未连接校园网或学校 VPN 时进入教务页就发起受限请求
-- 可点击考勤卡片右下角刷新图标手动读取最新数据，旁边会显示上次刷新时间
-- 可在设置页“自动刷新设置”中开启“体育查询自动刷新”并选择刷新间隔；开启后进入教务中心会主动读取一次，并按设置间隔刷新
-- 每次读取前都会重新执行校园网 / VPN 前置检测；不可访问受限校园站点时不提交体育部密码
-
-查询流程：
-
-- 先执行校园网 / VPN 前置检测；不可访问受限校园站点时不提交体育部密码
-- 使用安全设置页保存的学工号作为 `txtuser`，使用“体育部查询密码”作为 `txtpwd`
-- 固定学生身份 `dlljs=st`，并携带体育部登录页当前返回的 ASP.NET WebForms 隐藏字段
-- 登录成功后只读访问 `https://tygl.sspu.edu.cn/sportscore/stScore.aspx?item=1`
-- 页面使用 GBK / GB2312 编码，应用解码后解析四类次数与明细记录
-
-已知交互限制：
-
-- 体育部查询密码与 OA 密码不同，缺少体育部查询密码时不会尝试复用 OA 密码
-- 登录失败、会话不可用或页面结构变化时，教务中心页会展示明确错误状态
-- “查看考勤记录”会打开二级页面展示明细；仅展示只读结果，不执行次数调整等写入操作
-
-登录与解析规则探索记录见 [SPORTS_ATTENDANCE_RULES.md](SPORTS_ATTENDANCE_RULES.md)。
+每次读取前都会重新执行校园网 / VPN 前置检测，使用独立体育部查询密码，不复用 OA 密码。登录与解析规则探索记录见 [SPORTS_ATTENDANCE_RULES.md](SPORTS_ATTENDANCE_RULES.md)。
 
 ## 6.7 校园卡余额与交易记录调试
 
-主页提供“校园卡余额”卡片，展示当前账户余额、非正常卡状态提示和最近交易记录。
+主页提供“校园卡余额”卡片，展示余额、卡状态和最近交易记录；详情页支持按日期范围只读查询交易记录。默认不自动访问校园卡系统；可手动刷新，也可在设置页开启“校园卡余额自动刷新”。
 
-刷新方式：
-
-- 默认不自动访问校园卡系统，避免未连接校园网或学校 VPN 时进入主页就发起受限请求
-- 可点击卡片右下角刷新图标手动读取最新余额、卡状态和交易记录，旁边会显示上次刷新时间
-- 可在设置页“自动刷新设置”中开启“校园卡余额自动刷新”并选择刷新间隔；开启后进入主页会主动读取一次，并按设置间隔刷新
-- 每次读取前都会重新执行校园网 / VPN 前置检测；不可访问受限校园站点时不进入 OA / 校园卡登录链路
-
-查询流程：
-
-- 使用 `https://oa.sspu.edu.cn//interface/Entrance.jsp?id=xykxt` 作为 OA 校园卡入口
-- 复用已保存的 OA/CAS Cookie 会话；会话失效时调用现有 OA 登录校验刷新会话
-- OA 门户页会先脚本跳转到 `https://card.sspu.edu.cn?sysadmin=...`，业务页面实际继续进入 `http://card.sspu.edu.cn/epay/`
-- 校园卡业务入口按 `http://card.sspu.edu.cn/epay/`、`/epay/myepay/index`、`/epay/consume/index`、`/epay/consume/query` 这一组只读候选路径访问，CAS 回调服务为 `http://card.sspu.edu.cn/epay/j_spring_cas_security_check`
-- 读取余额 / 状态页面和交易记录页面后解析余额、卡状态、最近交易记录
-- 详情页支持按日期范围查询交易记录；该操作仅访问交易查询接口，不执行充值、支付或其它写入操作
-
-已知交互限制：
-
-- 缺少学工号或 OA 密码时不会访问校园卡系统
-- OA/CAS 要求验证码、MFA 或安全验证时，需要先在安全设置中完成 OA 登录校验
-- 页面结构变化时，主页和详情页会展示明确错误状态，不会伪造余额或交易记录
-
-登录与解析规则探索记录见 [CAMPUS_CARD_RULES.md](CAMPUS_CARD_RULES.md)。
+校园卡链路复用 OA/CAS 会话，每次读取前都会执行校园网 / VPN 前置检测，不提供充值、支付或其它写入入口。登录与解析规则探索记录见 [CAMPUS_CARD_RULES.md](CAMPUS_CARD_RULES.md)。
 
 ## 6.8 学校邮箱只读收信调试
 
-侧边导航提供“学校邮箱”页面，用于校验邮箱协议登录状态，并通过 IMAP 或 POP 读取最近邮件。
+侧边导航提供“学校邮箱”页面，用于校验邮箱协议登录状态，并通过 IMAP 或 POP 读取最近邮件。邮箱账号固定由学工号派生为 `学工号@sspu.edu.cn`，邮箱密码独立于 OA 密码。
 
-刷新方式：
-
-- 默认不自动访问邮箱系统，避免进入邮箱页即下载邮件正文
-- 可点击“读取最近邮件”手动读取，读取成功后会显示上次刷新时间
-- 可在设置页“自动刷新设置”中开启“学校邮箱自动刷新”并选择刷新间隔；开启后进入邮箱页会主动读取一次，并按设置间隔刷新
-- 邮箱系统不执行校园网 / VPN 前置检测，网络失败会直接在页面提示邮箱协议连接异常
-
-配置方式：
-
-- 在设置页“教务凭据”区域保存学工号和“邮箱密码”
-- 邮箱账号固定为“学工号@sspu.edu.cn”，应用会根据已保存学工号自动派生
-- 邮箱密码独立于 OA 密码；缺少邮箱密码时不会尝试复用 OA 密码
-
-读取与校验流程：
-
-- IMAP 默认访问 `imap.exmail.qq.com:993`，使用 TLS，并通过 `EXAMINE INBOX` + `BODY.PEEK[]` 读取最近邮件
-- POP 默认访问 `pop.exmail.qq.com:995`，使用 TLS，通过 `LIST` + `RETR` 读取最近邮件，不调用删除命令
-- SMTP 默认访问 `smtp.exmail.qq.com:465`，仅执行 `EHLO` + `AUTH` 校验，不发送邮件
-
-只读限制：
-
-- 页面仅展示标题、发件人、时间、正文摘要和正文快照
-- 不提供发送、回复、转发、删除、移动、标记已读或修改文件夹入口
-- 错误信息不会展示邮箱密码、认证载荷或服务器会话值
-
-登录与协议规则探索记录见 [EMAIL_RULES.md](EMAIL_RULES.md)。
+页面仅展示标题、发件人、时间、正文摘要和正文快照；SMTP 仅用于认证校验，不提供发送、回复、转发、删除、移动或标记已读入口。登录与协议规则探索记录见 [EMAIL_RULES.md](EMAIL_RULES.md)。
 
 ## 6.9 学工报表第二课堂学分调试
 
-教务中心页提供“第二课堂学分”卡片，用于通过学工报表系统只读读取第二课堂逐项得分明细。卡片展示得分记录数量和部分项目得分，不再将单项分值相加为总学分。
+教务中心页提供“第二课堂学分”卡片，通过学工报表系统只读读取第二课堂逐项得分明细，展示得分记录数量和部分项目得分，不再将单项分值相加为总学分。
 
-刷新方式：
-
-- 默认不自动访问学工报表系统，避免未连接校园网或学校 VPN 时进入教务中心就发起受限请求
-- 可点击卡片右上角刷新图标手动读取最新学分，卡片右下角显示上次刷新时间
-- 可在设置页“自动刷新设置”中开启“第二课堂学分自动刷新”并选择刷新间隔；开启后进入教务中心会主动读取一次，并按设置间隔刷新
-- 每次读取前都会重新执行校园网 / VPN 前置检测；不可访问受限校园站点时不进入 OA / 学工报表登录链路
-
-查询流程：
-
-- 使用 `https://oa.sspu.edu.cn/interface/Entrance.jsp?id=xgreport` 作为 OA 学工报表入口
-- 复用已保存的 OA/CAS Cookie 会话；会话失效时调用现有 OA 登录校验刷新会话
-- 读取 `https://xgbb.sspu.edu.cn/sharedc/core/home/index.do` 后定位“第二课堂学分查询”入口
-- 只读打开第二课堂学分页面，解析类别、项目名称、认定时间、状态和学分字段
-
-已知交互限制：
-
-- 缺少学工号或 OA 密码时不会访问学工报表系统
-- `xgbb` 本地登录页包含验证码，应用不会提交该表单，只通过 OA/CAS 入口复用登录态
-- OA/CAS 要求验证码、MFA 或安全验证时，需要先在安全设置中完成 OA 登录校验
-- 页面结构变化时，教务中心页会展示明确错误状态，不会伪造学分或空成功结果
-
-登录与解析规则探索记录见 [STUDENT_REPORT_RULES.md](STUDENT_REPORT_RULES.md)。
+学工报表链路复用 OA/CAS 会话，每次读取前都会执行校园网 / VPN 前置检测；本地验证码登录页不会被提交。登录与解析规则探索记录见 [STUDENT_REPORT_RULES.md](STUDENT_REPORT_RULES.md)。
 
 ## 6.10 本专科教务只读摘要与独立课程表调试
 
-教务中心页新增“本专科教务”摘要卡片，课程表能力则通过独立“课表”页面展示。两者共用同一套 EAMS 只读服务与自动刷新配置。
+教务中心页新增“本专科教务”摘要卡片，独立“课表”页面展示当前学期课表。两者共用 EAMS 只读服务与自动刷新配置，默认不自动访问本专科教务系统。
 
-刷新方式：
-
-- 默认不自动访问本专科教务系统，避免未连接校园网或学校 VPN 时进入页面即发起受限请求
-- 可点击教务中心摘要卡片右上角刷新图标手动读取当前摘要
-- 可点击侧边导航“课表”进入独立课程表页面，并通过“刷新课表”按钮只读拉取当前学期课程
-- 可在设置页“自动刷新设置”中开启“本专科教务自动刷新”并选择刷新间隔；开启后进入教务中心或课程表页会主动读取一次，并按设置间隔刷新
-- 每次读取前都会重新执行校园网 / VPN 前置检测；不可访问受限校园站点时不进入 OA / EAMS 登录链路
-
-查询流程：
-
-- 使用 `https://oa.sspu.edu.cn/interface/Entrance.jsp?id=bzkjw` 作为 OA 本专科教务入口
-- 复用已保存的 OA/CAS Cookie 会话；会话失效时调用现有 OA 登录校验刷新会话
-- 业务站点访问 `https://jx.sspu.edu.cn/eams/` 时固定使用 Firefox UA，降低 EAMS 页面兼容性差异带来的误判
-- 进入 EAMS 首页后按只读方式识别课表、成绩、考试、培养计划、开课检索和空闲教室入口；缺失入口或页面结构变化时会降级展示
-- 培养计划完成情况通过培养计划课程与已通过成绩做关联统计，不执行计划调整、课程替代申请或确认操作
-
-只读限制：
-
-- 页面只展示个人基本信息、课程名称、时间、地点、教师、周次、成绩、考试时间地点、培养计划和入口状态
-- 不提供选课、退课、调课、教学评价、确认、提交申请、预约教室或任何写入入口
-- 错误信息不会展示 OA 密码、Cookie、CAS Ticket、Token 或个人详细业务数据
-
-已知限制：
-
-- 开课检索与空闲教室当前已在服务层按只读表单方式支持，并在摘要卡片中展示入口识别状态；后续如需更完整的交互式筛选页面，可在此基础上继续扩展
-- 若 EAMS 页面结构调整，摘要卡片和课程表页会展示明确错误状态，不会伪造课表、成绩或考试结果
+本专科教务固定使用 `https://oa.sspu.edu.cn/interface/Entrance.jsp?id=bzkjw` 入口并复用 OA/CAS 会话，业务站点使用 Firefox UA。页面仅展示个人基本信息、课表、成绩、考试、培养计划和入口状态，不提供选课、退课、调课、教学评价、确认、提交申请、预约教室或任何写入入口。
 
 ## 7. 构建发布包
