+++ lib/widgets/settings_auto_refresh_section.dart
@@ -11,6 +11,8 @@ import 'package:fluent_ui/fluent_ui.dart';
 import '../theme/fluent_tokens.dart';
 import 'settings_widgets.dart';
 
+part 'settings_auto_refresh_rows.dart';
+
 /// 设置页自动刷新设置分区。
 class SettingsAutoRefreshSection extends StatelessWidget {
   /// 校园网 / VPN 状态检测间隔，单位分钟。
@@ -239,292 +241,6 @@ class SettingsAutoRefreshSection extends StatelessWidget {
             onCampusNetworkDetectionIntervalChanged(value);
           }
         },
-      ),
-    );
-  }
-
-  Widget _buildSportsAttendanceAutoRefreshRow(BuildContext context) {
-    final theme = FluentTheme.of(context);
-    return buildResponsiveSettingsRow(
-      context: context,
-      icon: FluentIcons.running,
-      title: Text('体育查询自动刷新', style: theme.typography.bodyStrong),
-      subtitle: Text(
-        '控制教务中心课外活动考勤卡片的自动读取；体育查询需要校园网或学校 VPN，关闭后仍可在卡片右上角手动刷新',
-        style: theme.typography.caption,
-      ),
-      trailing: Wrap(
-        spacing: FluentSpacing.s,
-        runSpacing: FluentSpacing.s,
-        crossAxisAlignment: WrapCrossAlignment.center,
-        children: [
-          ToggleSwitch(
-            checked: sportsAttendanceAutoRefreshEnabled,
-            onChanged: (value) => onSportsAttendanceAutoRefreshChanged(value),
-          ),
-          _buildSportsAttendanceIntervalComboBox(),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSportsAttendanceIntervalComboBox() {
-    final enabledIntervalOptions = Map<int, String>.fromEntries(
-      kIntervalOptions.entries.where((entry) => entry.key > 0),
-    );
-    final selectedValue =
-        enabledIntervalOptions.containsKey(
-          sportsAttendanceAutoRefreshIntervalMinutes,
-        )
-        ? sportsAttendanceAutoRefreshIntervalMinutes
-        : 30;
-
-    return ConstrainedBox(
-      constraints: const BoxConstraints(maxWidth: 180),
-      child: ComboBox<int>(
-        isExpanded: true,
-        value: selectedValue,
-        items: enabledIntervalOptions.entries
-            .map(
-              (entry) =>
-                  ComboBoxItem<int>(value: entry.key, child: Text(entry.value)),
-            )
-            .toList(),
-        onChanged: sportsAttendanceAutoRefreshEnabled
-            ? (value) {
-                if (value != null) {
-                  onSportsAttendanceAutoRefreshIntervalChanged(value);
-                }
-              }
-            : null,
-      ),
-    );
-  }
-
-  Widget _buildCampusCardAutoRefreshRow(BuildContext context) {
-    final theme = FluentTheme.of(context);
-    return buildResponsiveSettingsRow(
-      context: context,
-      icon: FluentIcons.payment_card,
-      title: Text('校园卡余额自动刷新', style: theme.typography.bodyStrong),
-      subtitle: Text(
-        '控制主页校园卡余额卡片的自动读取；需要校园网或学校 VPN 与 OA 登录，关闭后仍可在卡片右下角手动刷新',
-        style: theme.typography.caption,
-      ),
-      trailing: Wrap(
-        spacing: FluentSpacing.s,
-        runSpacing: FluentSpacing.s,
-        crossAxisAlignment: WrapCrossAlignment.center,
-        children: [
-          ToggleSwitch(
-            checked: campusCardAutoRefreshEnabled,
-            onChanged: (value) => onCampusCardAutoRefreshChanged(value),
-          ),
-          _buildCampusCardIntervalComboBox(),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildCampusCardIntervalComboBox() {
-    final enabledIntervalOptions = Map<int, String>.fromEntries(
-      kIntervalOptions.entries.where((entry) => entry.key > 0),
-    );
-    final selectedValue =
-        enabledIntervalOptions.containsKey(campusCardAutoRefreshIntervalMinutes)
-        ? campusCardAutoRefreshIntervalMinutes
-        : 30;
-
-    return ConstrainedBox(
-      constraints: const BoxConstraints(maxWidth: 180),
-      child: ComboBox<int>(
-        isExpanded: true,
-        value: selectedValue,
-        items: enabledIntervalOptions.entries
-            .map(
-              (entry) =>
-                  ComboBoxItem<int>(value: entry.key, child: Text(entry.value)),
-            )
-            .toList(),
-        onChanged: campusCardAutoRefreshEnabled
-            ? (value) {
-                if (value != null) {
-                  onCampusCardAutoRefreshIntervalChanged(value);
-                }
-              }
-            : null,
-      ),
-    );
-  }
-
-  Widget _buildEmailAutoRefreshRow(BuildContext context) {
-    final theme = FluentTheme.of(context);
-    return buildResponsiveSettingsRow(
-      context: context,
-      icon: FluentIcons.mail,
-      title: Text('学校邮箱自动刷新', style: theme.typography.bodyStrong),
-      subtitle: Text(
-        '控制学校邮箱页面的自动收信；邮箱系统不要求校园网或 VPN，关闭后仍可在邮箱页手动读取',
-        style: theme.typography.caption,
-      ),
-      trailing: Wrap(
-        spacing: FluentSpacing.s,
-        runSpacing: FluentSpacing.s,
-        crossAxisAlignment: WrapCrossAlignment.center,
-        children: [
-          ToggleSwitch(
-            checked: emailAutoRefreshEnabled,
-            onChanged: (value) => onEmailAutoRefreshChanged(value),
-          ),
-          _buildEmailIntervalComboBox(),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildEmailIntervalComboBox() {
-    final enabledIntervalOptions = Map<int, String>.fromEntries(
-      kIntervalOptions.entries.where((entry) => entry.key > 0),
-    );
-    final selectedValue =
-        enabledIntervalOptions.containsKey(emailAutoRefreshIntervalMinutes)
-        ? emailAutoRefreshIntervalMinutes
-        : 30;
-
-    return ConstrainedBox(
-      constraints: const BoxConstraints(maxWidth: 180),
-      child: ComboBox<int>(
-        isExpanded: true,
-        value: selectedValue,
-        items: enabledIntervalOptions.entries
-            .map(
-              (entry) =>
-                  ComboBoxItem<int>(value: entry.key, child: Text(entry.value)),
-            )
-            .toList(),
-        onChanged: emailAutoRefreshEnabled
-            ? (value) {
-                if (value != null) {
-                  onEmailAutoRefreshIntervalChanged(value);
-                }
-              }
-            : null,
-      ),
-    );
-  }
-
-  Widget _buildStudentReportAutoRefreshRow(BuildContext context) {
-    final theme = FluentTheme.of(context);
-    return buildResponsiveSettingsRow(
-      context: context,
-      icon: FluentIcons.education,
-      title: Text('第二课堂学分自动刷新', style: theme.typography.bodyStrong),
-      subtitle: Text(
-        '控制教务中心第二课堂学分卡片的自动读取；需要校园网或学校 VPN 与 OA 登录，关闭后仍可在卡片右上角手动刷新',
-        style: theme.typography.caption,
-      ),
-      trailing: Wrap(
-        spacing: FluentSpacing.s,
-        runSpacing: FluentSpacing.s,
-        crossAxisAlignment: WrapCrossAlignment.center,
-        children: [
-          ToggleSwitch(
-            checked: studentReportAutoRefreshEnabled,
-            onChanged: (value) => onStudentReportAutoRefreshChanged(value),
-          ),
-          _buildStudentReportIntervalComboBox(),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildStudentReportIntervalComboBox() {
-    final enabledIntervalOptions = Map<int, String>.fromEntries(
-      kIntervalOptions.entries.where((entry) => entry.key > 0),
-    );
-    final selectedValue =
-        enabledIntervalOptions.containsKey(
-          studentReportAutoRefreshIntervalMinutes,
-        )
-        ? studentReportAutoRefreshIntervalMinutes
-        : 30;
-
-    return ConstrainedBox(
-      constraints: const BoxConstraints(maxWidth: 180),
-      child: ComboBox<int>(
-        isExpanded: true,
-        value: selectedValue,
-        items: enabledIntervalOptions.entries
-            .map(
-              (entry) =>
-                  ComboBoxItem<int>(value: entry.key, child: Text(entry.value)),
-            )
-            .toList(),
-        onChanged: studentReportAutoRefreshEnabled
-            ? (value) {
-                if (value != null) {
-                  onStudentReportAutoRefreshIntervalChanged(value);
-                }
-              }
-            : null,
-      ),
-    );
-  }
-
-  Widget _buildAcademicEamsAutoRefreshRow(BuildContext context) {
-    final theme = FluentTheme.of(context);
-    return buildResponsiveSettingsRow(
-      context: context,
-      icon: FluentIcons.education,
-      title: Text('本专科教务自动刷新', style: theme.typography.bodyStrong),
-      subtitle: Text(
-        '控制教务中心本专科教务摘要和独立课程表页面的自动读取；需要校园网或学校 VPN 与 OA 登录，关闭后仍可在页面中手动刷新',
-        style: theme.typography.caption,
-      ),
-      trailing: Wrap(
-        spacing: FluentSpacing.s,
-        runSpacing: FluentSpacing.s,
-        crossAxisAlignment: WrapCrossAlignment.center,
-        children: [
-          ToggleSwitch(
-            checked: academicEamsAutoRefreshEnabled,
-            onChanged: (value) => onAcademicEamsAutoRefreshChanged(value),
-          ),
-          _buildAcademicEamsIntervalComboBox(),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildAcademicEamsIntervalComboBox() {
-    final enabledIntervalOptions = Map<int, String>.fromEntries(
-      kIntervalOptions.entries.where((entry) => entry.key > 0),
-    );
-    final selectedValue =
-        enabledIntervalOptions.containsKey(
-          academicEamsAutoRefreshIntervalMinutes,
-        )
-        ? academicEamsAutoRefreshIntervalMinutes
-        : 30;
-
-    return ConstrainedBox(
-      constraints: const BoxConstraints(maxWidth: 180),
-      child: ComboBox<int>(
-        isExpanded: true,
-        value: selectedValue,
-        items: enabledIntervalOptions.entries
-            .map(
-              (entry) =>
-                  ComboBoxItem<int>(value: entry.key, child: Text(entry.value)),
-            )
-            .toList(),
-        onChanged: academicEamsAutoRefreshEnabled
-            ? (value) {
-                if (value != null) {
-                  onAcademicEamsAutoRefreshIntervalChanged(value);
-                }
-              }
-            : null,
       ),
     );
   }
